facenet2/requirements.txt
pillow
tensorflow
numpy
matplotlib